fix: improve code based on code review
Remove `__pycache__`  directories from everywhere please. Introduce .gitignore file and add these directories to that.
`PaymentUsecase` could be extracted to a pytest fixture in `/test/unit/usecase/test_payment_usecase.py`
Is `EmailOutputBoundary` meant to be a protocol? If yes, then, extend from `Protocol`. Looking at the InputBoundary which is a Protocol, I feel EmailBoundary could be defined as such.
My preferred naming for Repositories is `CreditRepository` this means that if you add more repos then every one of those has it’s own protocol, then you have a concrete version which can be based on technology or storage mechanism. For example, `InMemoryCreditRepository`, `MySQLCreditRepository` or `DynamoDbCreditRepository`
While this is ok in a small app, it’s worth doing in a larger application.